requirements.txt
astronomer-cosmos==1.10.1
dbt-snowflake
apache-airflow-providers-snowflake==5.5.0
snowflake-connector-python
pandas
